+++ src/NetworkTest/testQuality/helpers/filterDevicesForType.ts
@@ -1,13 +1,12 @@
 import * as Promise from 'promise';
 import * as e from '../../testConnectivity/errors';
 import { OT } from '../../types/opentok';
-import { OTError } from '../../types/opentok/error';
 
 export type InputDeviceType = 'audioInput' | 'videoInput';
 
 export default function filterDevicesForType(OT: OT.Client, type: InputDeviceType) {
   return new Promise((resolve, reject) => {
-    OT.getDevices((error?: OTError, devices: OT.Device[] = []) => {
+    OT.getDevices((error?: OT.Error, devices: OT.Device[] = []) => {
       if (error) {
         reject(new e.FailedToObtainMediaDevices());
       } else {
