+++ src/NetworkTest/testQuality/helpers/getQualityEvaluation.ts
@@ -1,7 +1,6 @@
 import config from './config';
 import { get } from '../../../util';
-import { AV } from '../types';
-import { AverageStatsBase } from './stats';
+import { AV, AverageStatsBase } from '../types/stats';
 
 export interface QualityEvaluationResults{
   supported: boolean;
